requirements.txt
streamlit
PyMuPDF
pandas
google-generativeai
git+https://github.com/<username>/langextract.git

